+++ docs/app/guides/get-started/installation/page.tsx
@@ -1,11 +1,14 @@
-"use client"
-
 import Code from "../../../../components/Code"
 import CodeAndExample from "../../../../components/CodeAndExample"
 
 import Plain from "../../../../components/guide-examples/hello-fx/plain"
 import PlainSource from "!raw-loader!../../../../components/guide-examples/hello-fx/plain"
 import endent from "endent"
+import type { Metadata } from "next"
+
+export const metadata: Metadata = {
+  title: "Installation",
+}
 
 function Page() {
   return (
