--- docs/app/head.tsx
@@ -1,10 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>Mafs: React components for interactive math</title>
-      <meta content="width=device-width, initial-scale=1" name="viewport" />
-      <meta name="description" content="React components for interactive math" />
-      <link rel="icon" href={"/favicon.svg"} />
-    </>
-  )
-}
