+++ docs/package.json
@@ -20,7 +20,7 @@
     "js-easing-functions": "^1.0.3",
     "lodash": "^4.17.21",
     "mafs": "link:../src",
-    "next": "^13.4.12",
+    "next": "^14.0.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-markdown": "^8.0.7",
